requirements.txt
torch
torchvision
numpy
pillow
nltk
tqdm
matplotlib
pyyaml
pycocoevalcap
